Use SVG icon for Tic Tac Toe
