+++ components/apps/tic-tac-toe.js
@@ -1,0 +1,73 @@
+import React, { useState } from 'react';
+
+const winningLines = [
+  [0, 1, 2],
+  [3, 4, 5],
+  [6, 7, 8],
+  [0, 3, 6],
+  [1, 4, 7],
+  [2, 5, 8],
+  [0, 4, 8],
+  [2, 4, 6],
+];
+
+const calculateWinner = (board) => {
+  for (const [a, b, c] of winningLines) {
+    if (board[a] && board[a] === board[b] && board[a] === board[c]) {
+      return board[a];
+    }
+  }
+  return null;
+};
+
+const TicTacToe = () => {
+  const [board, setBoard] = useState(Array(9).fill(null));
+  const [xIsNext, setXIsNext] = useState(true);
+
+  const winner = calculateWinner(board);
+  const isDraw = !winner && board.every(Boolean);
+  const status = winner
+    ? `${winner} wins!`
+    : isDraw
+    ? "It's a draw"
+    : `Next player: ${xIsNext ? 'X' : 'O'}`;
+
+  const handleClick = (index) => {
+    if (board[index] || winner) return;
+    const newBoard = board.slice();
+    newBoard[index] = xIsNext ? 'X' : 'O';
+    setBoard(newBoard);
+    setXIsNext(!xIsNext);
+  };
+
+  const reset = () => {
+    setBoard(Array(9).fill(null));
+    setXIsNext(true);
+  };
+
+  return (
+    <div className="h-full w-full flex flex-col items-center justify-center bg-ub-cool-grey text-white p-4">
+      <div className="grid grid-cols-3 gap-1 w-60 mb-4">
+        {board.map((cell, idx) => (
+          <button
+            key={idx}
+            className="h-20 w-20 bg-gray-700 hover:bg-gray-600 text-4xl flex items-center justify-center"
+            onClick={() => handleClick(idx)}
+          >
+            {cell}
+          </button>
+        ))}
+      </div>
+      <div className="mb-4">{status}</div>
+      <button
+        className="px-4 py-2 bg-gray-700 hover:bg-gray-600 rounded"
+        onClick={reset}
+      >
+        Reset
+      </button>
+    </div>
+  );
+};
+
+export default TicTacToe;
+
